Add quote API layer.

+++ src/modules/quote/application/usecase/remove-quote.uc.ts
@@ -4,10 +4,10 @@ import { RemoveQuoteDTO } from "../dto/quote.dto";
 export class RemoveQuoteUC {
   constructor(private quoteRepo: IQuoteRepo) {}
 
-  async execute({id}: RemoveQuoteDTO): Promise<void> {
+  async execute({id, userId}: RemoveQuoteDTO): Promise<void> {
     const quote = await this.quoteRepo.findById(id);
     if (!quote) throw new Error("Quote not found");
-
+    if (quote.userId !== userId) throw new Error("Unauthorized to remove this quote");
     await this.quoteRepo.remove(id);
   }
 }
